+++ src/Traits/Request/Patch.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+namespace Firehed\API\Traits\Request;
+
+use Firehed\API\Enums\HTTPMethod;
+
+trait Patch
+{
+    public function getMethod(): HTTPMethod
+    {
+        return HTTPMethod::PATCH();
+    }
+}